requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Cors>=4.0
python-dotenv>=1.0
twilio>=9.0
Werkzeug>=3.0
SQLAlchemy>=2.0
Werkzeug>=3.0

flask-jwt-extended>=4.6
twilio>=9.0
python-dotenv>=1.0